chore: split vike plugins
