+++ examples/demo/package.json
@@ -23,7 +23,7 @@
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
     "typescript": "^5.8.2",
-    "vike": "^0.4.228",
+    "vike": "^0.4.235",
     "vike-react": "^0.5.13",
     "vike-vercel": "workspace:^",
     "vite": "^7.0.0-beta.1",
