+++ packages/vike/package.json
@@ -37,7 +37,7 @@
   "dependencies": {
     "@brillout/libassert": "^0.5.8",
     "@photonjs/core": "0.0.1-commit-5f3addf",
-    "vike": "^0.4.228",
+    "vike": "^0.4.235",
     "vike-server": "1.0.15-commit-3ad9b16"
   }
 }
